feat: 자동차 구현 및 상수 선언
